+++ sim/rogue/backstab.go
@@ -54,12 +54,10 @@ func (rogue *Rogue) registerBackstabSpell() {
 
 			if result.Landed() {
 				rogue.AddComboPoints(sim, 1, spell.ComboPointMetrics())
-				// FIXME: Extension of a Rupture Dot can occur up to 3 times
-				ruptureDot := rogue.Rupture.Dot(target)
-				if hasGlyph && ruptureDot.IsActive() {
-					ruptureDot.NumberOfTicks += 1
-					ruptureDot.RecomputeAuraDuration()
-					ruptureDot.UpdateExpires(ruptureDot.ExpiresAt() + ruptureDot.TickLength)
+				if dot := rogue.Rupture.Dot(target); hasGlyph && dot.IsActive() && dot.NumberOfTicks < dot.MaxStacks+3 {
+					dot.NumberOfTicks += 1
+					dot.RecomputeAuraDuration()
+					dot.UpdateExpires(dot.ExpiresAt() + dot.TickLength)
 				}
 			} else {
 				spell.IssueRefund(sim)
