+++ sim/rogue/rupture.go
@@ -12,6 +12,7 @@ const RuptureSpellID = 48672
 
 func (rogue *Rogue) registerRupture() {
 	glyphTicks := core.TernaryInt32(rogue.HasMajorGlyph(proto.RogueMajorGlyph_GlyphOfRupture), 2, 0)
+
 	rogue.Rupture = rogue.RegisterSpell(core.SpellConfig{
 		ActionID:     core.ActionID{SpellID: RuptureSpellID},
 		SpellSchool:  core.SpellSchoolPhysical,
@@ -72,11 +73,11 @@ func (rogue *Rogue) registerRupture() {
 		ApplyEffects: func(sim *core.Simulation, target *core.Unit, spell *core.Spell) {
 			result := spell.CalcOutcome(sim, target, spell.OutcomeMeleeSpecialHit)
 			if result.Landed() {
-				comboPoints := rogue.ComboPoints()
+				numberOfTicks := 3 + rogue.ComboPoints() + glyphTicks
 				dot := spell.Dot(target)
 				dot.Spell = spell
-				dot.NumberOfTicks = 3 + comboPoints + glyphTicks
-				dot.RecomputeAuraDuration()
+				dot.NumberOfTicks = numberOfTicks
+				dot.MaxStacks = numberOfTicks // slightly hacky; used to determine max extra ticks from Glyph of Backstab
 				dot.Apply(sim)
 				rogue.ApplyFinisher(sim, spell)
 			} else {
@@ -87,8 +88,10 @@ func (rogue *Rogue) registerRupture() {
 	})
 }
 
+func (rogue *Rogue) RuptureTicks(comboPoints int32) int32 {
+	return 3 + comboPoints + core.TernaryInt32(rogue.HasMajorGlyph(proto.RogueMajorGlyph_GlyphOfRupture), 2, 0)
+}
+
 func (rogue *Rogue) RuptureDuration(comboPoints int32) time.Duration {
-	return time.Second*6 +
-		time.Second*2*time.Duration(comboPoints) +
-		core.TernaryDuration(rogue.HasMajorGlyph(proto.RogueMajorGlyph_GlyphOfRupture), time.Second*4, 0)
+	return time.Duration(rogue.RuptureTicks(comboPoints)) * time.Second * 2
 }
