+++ sim/rogue/subtlety_rotation.go
@@ -1,62 +1,25 @@
 package rogue
 
 import (
+	"log"
 	"time"
 
 	"github.com/wowsims/wotlk/sim/core"
 	"github.com/wowsims/wotlk/sim/core/proto"
 )
 
-type subtletyPrio struct {
-	check GetAction
-	cast  DoAction
-	cost  float64
+type subtlety_rotation struct {
+	prios []prio
 }
 
-func (rogue *Rogue) setSubtletyBuilder(sim *core.Simulation) {
-	mhDagger := rogue.Equip[proto.ItemSlot_ItemSlotMainHand].WeaponType == proto.WeaponType_WeaponTypeDagger
-	// Garrote
-	if !rogue.Garrote.CurDot().Aura.IsActive() && rogue.ShadowDanceAura.IsActive() && !rogue.PseudoStats.InFrontOfTarget {
-		rogue.Builder = rogue.Garrote
-		rogue.BuilderPoints = 1
-		return
-	}
-	// Ambush
-	if rogue.ShadowDanceAura.IsActive() && mhDagger && !rogue.PseudoStats.InFrontOfTarget {
-		rogue.Builder = rogue.Ambush
-		rogue.BuilderPoints = 2
-		return
-	}
-	// Backstab
-	if mhDagger && !rogue.Rotation.HemoWithDagger && !rogue.PseudoStats.InFrontOfTarget {
-		rogue.Builder = rogue.Backstab
-		rogue.BuilderPoints = 1
-		return
-	}
-	// Ghostly Strike
-	if rogue.Talents.GhostlyStrike && rogue.Rotation.UseGhostlyStrike && rogue.GhostlyStrike.IsReady(sim) {
-		rogue.Builder = rogue.GhostlyStrike
-		rogue.BuilderPoints = 1
-		return
-	}
-	// Hemorrhage
-	if rogue.Talents.Hemorrhage {
-		rogue.Builder = rogue.Hemorrhage
-		rogue.BuilderPoints = 1
-	} else
-	// Sinister Strike
-	{
-		rogue.Builder = rogue.SinisterStrike
-		rogue.BuilderPoints = 1
-	}
-}
+func (x *subtlety_rotation) setup(sim *core.Simulation, rogue *Rogue) {
+	x.setSubtletyBuilder(sim, rogue)
 
-func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
-	rogue.subtletyPrios = rogue.subtletyPrios[:0]
+	x.prios = x.prios[:0]
 
 	if rogue.Rotation.OpenWithPremeditation && rogue.Talents.Premeditation {
 		hasCastPremeditation := false
-		rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+		x.prios = append(x.prios, prio{
 			func(s *core.Simulation, r *Rogue) PriorityAction {
 				if hasCastPremeditation {
 					return Skip
@@ -79,7 +42,7 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 
 	if rogue.Rotation.OpenWithShadowstep && rogue.Talents.Shadowstep {
 		hasCastShadowstep := false
-		rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+		x.prios = append(x.prios, prio{
 			func(s *core.Simulation, r *Rogue) PriorityAction {
 				if hasCastShadowstep {
 					return Skip
@@ -103,7 +66,7 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 	// Garrote
 	if rogue.Rotation.OpenWithGarrote {
 		hasCastGarrote := false
-		rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+		x.prios = append(x.prios, prio{
 			func(sim *core.Simulation, rogue *Rogue) PriorityAction {
 				if hasCastGarrote {
 					return Skip
@@ -125,7 +88,7 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 	}
 
 	// Slice and Dice
-	rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+	x.prios = append(x.prios, prio{
 		func(sim *core.Simulation, rogue *Rogue) PriorityAction {
 			if rogue.SliceAndDiceAura.IsActive() {
 				return Skip
@@ -155,7 +118,7 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 	if rogue.Rotation.ExposeArmorFrequency == proto.Rogue_Rotation_Once ||
 		rogue.Rotation.ExposeArmorFrequency == proto.Rogue_Rotation_Maintain {
 		hasCastExpose := false
-		rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+		x.prios = append(x.prios, prio{
 			func(sim *core.Simulation, rogue *Rogue) PriorityAction {
 				if hasCastExpose && rogue.Rotation.ExposeArmorFrequency == proto.Rogue_Rotation_Once {
 					return Skip
@@ -208,7 +171,7 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 	}
 
 	// Enable CDS
-	rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+	x.prios = append(x.prios, prio{
 		func(sim *core.Simulation, rogue *Rogue) PriorityAction {
 			if rogue.allMCDsDisabled {
 				for _, mcd := range rogue.GetMajorCooldowns() {
@@ -226,7 +189,7 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 
 	//Shadowstep
 	if rogue.Talents.Shadowstep {
-		rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+		x.prios = append(x.prios, prio{
 			func(sim *core.Simulation, rogue *Rogue) PriorityAction {
 				if rogue.Shadowstep.IsReady(sim) {
 					// Can we cast Rupture now?
@@ -248,7 +211,7 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 	const ruptureMinDuration = time.Second * 8 // heuristically, 3-4 Rupture ticks are better DPE than Eviscerate or Envenom
 
 	// Rupture
-	rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+	x.prios = append(x.prios, prio{
 		func(sim *core.Simulation, rogue *Rogue) PriorityAction {
 			if rogue.Rupture.CurDot().IsActive() || sim.GetRemainingDuration() < ruptureMinDuration {
 				return Skip
@@ -270,7 +233,7 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 
 	//Envenom
 	if rogue.Rotation.SubtletyFinisherPriority == proto.Rogue_Rotation_SubtletyEnvenom {
-		rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+		x.prios = append(x.prios, prio{
 			func(sim *core.Simulation, rogue *Rogue) PriorityAction {
 				if !rogue.DeadlyPoison.CurDot().Aura.IsActive() {
 					return Skip
@@ -294,7 +257,7 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 	}
 
 	// Eviscerate
-	rogue.subtletyPrios = append(rogue.subtletyPrios, subtletyPrio{
+	x.prios = append(x.prios, prio{
 		func(sim *core.Simulation, rogue *Rogue) PriorityAction {
 			// end of combat handling - prefer Eviscerate over Builder, heuristically
 			if sim.GetRemainingDuration().Seconds() < rogue.getExpectedSecondsPerComboPoint() && rogue.ComboPoints() >= 2 && rogue.CurrentEnergy() >= rogue.Eviscerate.DefaultCast.Cost {
@@ -315,10 +278,14 @@ func (rogue *Rogue) setupSubtletyRotation(sim *core.Simulation) {
 	})
 }
 
-func (rogue *Rogue) doSubtletyRotation(sim *core.Simulation) {
+func (x *subtlety_rotation) run(sim *core.Simulation, rogue *Rogue) {
+	if !rogue.GCD.IsReady(sim) {
+		log.Panic("gcd not ready")
+	}
+
 	prioIndex := 0
-	for prioIndex < len(rogue.subtletyPrios) {
-		prio := rogue.subtletyPrios[prioIndex]
+	for prioIndex < len(x.prios) {
+		prio := x.prios[prioIndex]
 		switch prio.check(sim, rogue) {
 		case Skip:
 			prioIndex += 1
@@ -331,7 +298,7 @@ func (rogue *Rogue) doSubtletyRotation(sim *core.Simulation) {
 			}
 
 			if rogue.GCD.IsReady(sim) {
-				rogue.setSubtletyBuilder(sim)
+				x.setSubtletyBuilder(sim, rogue)
 				if !rogue.Builder.Cast(sim, rogue.CurrentTarget) {
 					rogue.WaitForEnergy(sim, rogue.Builder.DefaultCast.Cost)
 					return
@@ -356,8 +323,39 @@ func (rogue *Rogue) doSubtletyRotation(sim *core.Simulation) {
 	rogue.DoNothing()
 }
 
-func (rogue *Rogue) OnCanActSubtlety(sim *core.Simulation) {
-	if rogue.GCD.IsReady(sim) {
-		rogue.doSubtletyRotation(sim)
+func (x *subtlety_rotation) setSubtletyBuilder(sim *core.Simulation, rogue *Rogue) {
+	// Garrote
+	if !rogue.Garrote.CurDot().Aura.IsActive() && rogue.ShadowDanceAura.IsActive() && !rogue.PseudoStats.InFrontOfTarget {
+		rogue.Builder = rogue.Garrote
+		rogue.BuilderPoints = 1
+		return
 	}
+	// Ambush
+	if rogue.ShadowDanceAura.IsActive() && !rogue.PseudoStats.InFrontOfTarget && rogue.HasDagger(core.MainHand) {
+		rogue.Builder = rogue.Ambush
+		rogue.BuilderPoints = 2
+		return
+	}
+	// Backstab
+	if !rogue.Rotation.HemoWithDagger && !rogue.PseudoStats.InFrontOfTarget && rogue.HasDagger(core.MainHand) {
+		rogue.Builder = rogue.Backstab
+		rogue.BuilderPoints = 1
+		return
+	}
+	// Ghostly Strike
+	if rogue.Talents.GhostlyStrike && rogue.Rotation.UseGhostlyStrike && rogue.GhostlyStrike.IsReady(sim) {
+		rogue.Builder = rogue.GhostlyStrike
+		rogue.BuilderPoints = 1
+		return
+	}
+	// Hemorrhage
+	if rogue.Talents.Hemorrhage {
+		rogue.Builder = rogue.Hemorrhage
+		rogue.BuilderPoints = 1
+		return
+	}
+
+	// Sinister Strike
+	rogue.Builder = rogue.SinisterStrike
+	rogue.BuilderPoints = 1
 }
